got delete post by id working, thanks to a youtube video, also put server actions in their on file
